add toggle between chats
