+++ src/App.jsx
@@ -1,7 +1,6 @@
 import { useState } from "react";
-// import SideBar from "./component/SideBar";
 import ChatSection from "./component/ChatSection";
-import SideBarUpdated from "./component/SidebarUpdate";
+import SideBar from "./component/SideBar";
 
 export default function App() {
   const [isOpen, setIsOpen] = useState(true);
@@ -10,8 +9,7 @@ export default function App() {
   };
   return (
     <div className="h-screen flex bg-gray-300">
-      {/* <SideBar isOpen={isOpen} setIsOpen={setIsOpen} /> */}
-      <SideBarUpdated
+      <SideBar
         isOpen={isOpen}
         setIsOpen={setIsOpen}
         handleSideBarToggle={handleSideBarToggle}
