+++ src/component/ChatSection.jsx
@@ -60,7 +60,7 @@ const ChatSection = ({ isOpen, setIsOpen, handleSideBarToggle }) => {
   }, [messages]);
 
   return (
-    <div className={`ml-3 flex flex-col max-h-screen w-full`}>
+    <div className={`ml-3 flex flex-col max-h-screen w-full ${isOpen && 'hidden md:flex'}`}>
       <div className="flex justify-between items-center p-4 bg-[#F1F3F4] border-b-2 border-gray-300 rounded-t-3xl">
         <button className="cursor-pointer" onClick={handleSideBarToggle}>
           {isOpen ? (
